# 구현 | Transform을 한틱 안돌아서 방향 안나오던것 생성자에서 수정             DiagBeam 구현

+++ Project/Scripts/CProjectileScript.h
@@ -11,7 +11,7 @@ struct Proj_Struct
     int             Damage;
     int             MaxRange;
 
-    float           LifeCount;
+    int           LifeCount;
     float           LifeTime;
 
     bool            TargetTile;
